+++ spec/default_spec.rb
@@ -3,11 +3,9 @@ describe 'beyondcompare::default' do
     ChefSpec::ServerRunner.new do |node|
     end.converge(described_recipe)
   end
-
   it 'includes beyondcompare::install recipe' do
     expect(chef_run).to include_recipe 'beyondcompare::install'
   end
-
   it 'includes beyondcompre::gitconfig recipe' do
     expect(chef_run).to include_recipe 'beyondcompare::gitconfig'
   end