+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 This file is used to list changes made in each version of beyondcompare.
 
-## 0.0.10:
+## 1.0.0:
+- SuperMarket release for BeyondCompare 4.1.9
 
+## 0.0.10:
+- Initial SuperMarket release for BeyondCompare 3.3.10
 
 
 - - -